pyproject.toml
[project]
name = "wordbook"
version = "0.1.0"
description = ""
authors = [
    {name = "miho nakayama",email = "miho.nakayam@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,!=3.9.7,<4.0"
dependencies = [
    "notion-client>=2.2.1",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
]

[tool.poetry]
packages = [{include = "wordbook", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
